Use Google Fonts metadata to fetch all available variants
Previously, this theme retrieved a hardcoded list of font variants, as
specified by the mkdocs-material theme.  However, it was reported in #193
that some fonts do not have all of these variants, leading to an error
when trying to use such a font.

With this change, the precise list of variants for each font is used
by fetching the complete list of fonts and variants from the Google
Fonts API.

This commit also factors out the Google Fonts retrieval logic from
`external_resource_cache.py` into a separate `google_fonts.py` module.

Fixes #193.
