+++ sphinx_immaterial/__init__.py
@@ -249,6 +249,7 @@ def setup(app: Sphinx):
     app.connect("config-inited", _config_inited)
 
     app.setup_extension("sphinx_immaterial.external_resource_cache")
+    app.setup_extension("sphinx_immaterial.google_fonts")
 
     app.setup_extension(apidoc_formatting.__name__)
     app.setup_extension("sphinx_immaterial.apidoc.python.default")
